requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.2
django-celery-beat==2.8.1
django-redis==5.4.0
django-timezone-field==7.1
django_celery_results==2.6.0
flower==2.0.1
humanize==4.12.3
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
